Part 1 of the fixes for: Issue 301: configure number of particpents of meeting
in theory endless people can now access a meeting.
but recorded and i guess als *events* won't work now.
also the possibility of configuring in the backend the number of participants is miss for the moment.
added documentation to the basevideoobject and videocontainer
there is now a new attribute broadcastId, this is the new stream-name-id. The streamid is a unique ID for _each_ client. But a client _could_ use several broadCastID's to publish his setup.
I hope to fix it that way without reconnecting netconneciton, just be reseting the NetStream and publsih with a new broadCastID